+++ internal/provider/provider.go
@@ -189,6 +189,7 @@ func (p *IncidentProvider) DataSources(ctx context.Context) []func() datasource.
 		NewIncidentScheduleBetaDataSource,
 		NewIncidentScheduleRotationBetaDataSource,
 		NewIncidentIncidentTypesDataSource,
+		NewIncidentIncidentTemplateDataSource,
 		NewIncidentEscalationPathDataSource,
 		NewIncidentWorkflowDataSource,
 		NewRichTextDataSource,